refactor - move custom routers logic to config, subclassing for capapi and tracking_tool
